refactor: extract god-file responsibilities (Phase 9b, TD-1)
Behavior-preserving extractions along stable responsibility boundaries:
- server/index.ts: scanBaseDir + normalizeBaseDirs -> new server/list-scanner.ts
- cli/ingest.ts: collectFiles -> new cli/file-collection.ts
- parser/index.ts: extractPdfPages + PDF page types -> new parser/pdf-extract.ts
- base-dirs.ts: correct dedupAndPruneRoots comment (two-pass, not single-pass)

Net -319/+28 lines across the modified files; new modules are pure/cohesive.
Full per-tool-handler split of RAGServer was intentionally NOT done (high churn,
low correctness ROI; LOC is a warning signal not a hard target per TD-1).

check:all green, 837 tests pass.
