+++ src/cli/ingest.ts
@@ -1,17 +1,16 @@
 // CLI ingest subcommand — bulk file ingestion with single optimize() at end
 
-import { realpath, stat } from 'node:fs/promises'
-import { extname, resolve, sep } from 'node:path'
+import { stat } from 'node:fs/promises'
+import { resolve, sep } from 'node:path'
 
 import { SemanticChunker } from '../chunker/index.js'
 import type { Embedder } from '../embedder/index.js'
 import { buildChunksAndEmbeddings, buildVectorChunks } from '../ingest/compute.js'
 import { prepareVisualPdfChunks } from '../ingest/visual.js'
-import { DocumentParser, SUPPORTED_EXTENSIONS } from '../parser/index.js'
+import { DocumentParser } from '../parser/index.js'
 import type { QualityProfile } from '../pdf-visual/types.js'
 import type { BaseDirsConfig, BaseDirsConfigWarning } from '../utils/base-dirs.js'
-import { DEFAULT_MAX_FILE_SIZE, MAX_SCAN_DEPTH } from '../utils/limits.js'
-import { bfsCollectSupportedFiles } from '../utils/scan.js'
+import { DEFAULT_MAX_FILE_SIZE } from '../utils/limits.js'
 import type { VectorStore } from '../vectordb/index.js'
 import {
   createEmbedder,
@@ -19,6 +18,7 @@ import {
   resolveCliBaseDirsOrExit,
   toErrorMessage,
 } from './common.js'
+import { collectFiles } from './file-collection.js'
 import type { GlobalOptions, ResolvedGlobalConfig } from './options.js'
 import {
   consumeBaseDirArg,
@@ -276,69 +276,6 @@ export async function resolveConfig(
     resolved.chunkMinLength = chunkMinLength
   }
   return resolved
-}
-
-// ============================================
-// File Collection
-// ============================================
-
-async function collectFiles(
-  targetPath: string,
-  baseDirs: readonly string[],
-  excludePaths: string[]
-): Promise<string[]> {
-  const resolved = resolve(targetPath)
-  const info = await stat(resolved)
-
-  if (info.isFile()) {
-    const ext = extname(resolved).toLowerCase()
-    if (!SUPPORTED_EXTENSIONS.has(ext)) {
-      console.error(
-        `Unsupported file extension: ${ext} (supported: ${[...SUPPORTED_EXTENSIONS].join(', ')})`
-      )
-      return []
-    }
-    return [resolved]
-  }
-
-  if (info.isDirectory()) {
-    // realpath both sides so a symlinked positional path still matches a
-    // root whose realpath agrees. baseDirs from resolveCliBaseDirsOrExit
-    // are already realpath-normalized with a trailing sep.
-    const realResolved = await realpath(resolved)
-    const realResolvedWithSep = realResolved.endsWith(sep) ? realResolved : realResolved + sep
-    const insideAnyRoot = baseDirs.some(
-      (root) => realResolvedWithSep === root || realResolvedWithSep.startsWith(root)
-    )
-    if (!insideAnyRoot) {
-      console.error(
-        `Error: ${targetPath} is not under any configured base directory. ` +
-          `Allowed roots: ${baseDirs.join(', ')}. ` +
-          `Provide a path inside one of the configured roots, or set BASE_DIRS / --base-dir to include the desired tree.`
-      )
-      process.exit(1)
-    }
-
-    const {
-      files: collected,
-      unreadableDirs,
-      depthLimited,
-    } = await bfsCollectSupportedFiles(resolved, excludePaths)
-
-    for (const { dirPath } of unreadableDirs) {
-      console.error(`Warning: cannot read directory: ${dirPath}`)
-    }
-
-    if (depthLimited) {
-      console.error(
-        `Warning: some directories were skipped because they exceed the maximum depth (${MAX_SCAN_DEPTH})`
-      )
-    }
-
-    return [...new Set(collected)].sort()
-  }
-
-  return []
 }
 
 // ============================================
